refactor(workers): move timing constants to environment settings
- Add logging timing constants to core (`DEFAULT_LOG_STARTUP_S`, `DEFAULT_LOG_SUMMARY_INTERVAL_S`).
- Refactor BirdNET worker to use `POLLING_INTERVAL_S` and `DB_RETRY_INTERVAL_S` from `BirdnetEnvSettings` instead of hardcoded sleeps.
- Update Processor service to configure `UploadStats` durations via `ProcessorEnvSettings`.
- Change `UploadStats` signature to use `float` types for duration arguments.
- Update BirdNET worker tests to assert against dynamic environment settings.

+++ services/processor/src/silvasonic/processor/modules/upload_stats.py
@@ -17,7 +17,9 @@ log = structlog.get_logger()
 class UploadStats:
     """Tracks upload progress and limits log spam during steady state."""
 
-    def __init__(self, startup_duration_s: int = 300, summary_interval_s: int = 300) -> None:
+    def __init__(
+        self, startup_duration_s: float = 300.0, summary_interval_s: float = 300.0
+    ) -> None:
         """Initialize the stats tracker.
 
         Args:
